Fix #41 : add new files to decompose functions
